Path finder moved to separate folder
